re-organize code across multiple files
This seeks to make it easier to find funcs and units of functionality.
This also paves a path towards larger refactoring down the road.
