+++ src/helpers.go
@@ -17,8 +17,6 @@ import (
 	"fmt"
 	"net/http"
 	"os"
-	"strconv"
-	"strings"
 
 	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli"
@@ -33,92 +31,6 @@ func client(c *cli.Context) *vinyldns.Client {
 		Host:       c.GlobalString(hostFlag),
 		HTTPClient: &http.Client{},
 	}
-}
-
-func getGroup(c *vinyldns.Client, name, id string) (*vinyldns.Group, error) {
-	if name != "" {
-		return groupByName(c, name)
-	}
-
-	return c.Group(id)
-}
-
-func groupByName(c *vinyldns.Client, name string) (*vinyldns.Group, error) {
-	var g *vinyldns.Group
-	groups, err := c.Groups()
-	if err != nil {
-		return g, err
-	}
-
-	for _, group := range groups {
-		if group.Name == name {
-			return &group, nil
-		}
-	}
-
-	return g, fmt.Errorf("Group %s not found", name)
-}
-
-func getAdminGroupID(c *vinyldns.Client, id, name string) (string, error) {
-	if id != "" {
-		return id, nil
-	}
-
-	g, err := groupByName(c, name)
-	if err != nil {
-		return "", err
-	}
-
-	return g.ID, nil
-}
-
-func zoneByName(c *vinyldns.Client, name string) (vinyldns.Zone, error) {
-	var z vinyldns.Zone
-	zones, err := c.Zones()
-	if err != nil {
-		return z, err
-	}
-
-	for _, zone := range zones {
-		if zone.Name == name {
-			return zone, nil
-		}
-	}
-
-	return z, fmt.Errorf("Zone %s not found", name)
-}
-
-func getZoneID(c *vinyldns.Client, id, name string) (string, error) {
-	if id != "" {
-		return id, nil
-	}
-
-	z, err := zoneByName(c, name)
-	if err != nil {
-		return "", err
-	}
-
-	return z.ID, nil
-}
-
-func getZone(c *vinyldns.Client, name, id string) (vinyldns.Zone, error) {
-	if name != "" {
-		return zoneByName(c, name)
-	}
-
-	return c.Zone(id)
-}
-
-func typeSwitch(t string) string {
-	switch t {
-	case "A", "a":
-		return "A"
-	case "AAAA", "aaaa":
-		return "AAAA"
-	case "CNAME", "cname":
-		return "CNAME"
-	}
-	return ""
 }
 
 func getOption(c *cli.Context, name string) (string, error) {
@@ -154,58 +66,6 @@ func validateEnv(c *cli.Context) {
 	}
 }
 
-func getRecordValue(records []string, recordValue interface{}, recordPrepend string) []string {
-	var strVal string
-	switch recordValue.(type) {
-	case int:
-		strVal = strconv.Itoa(recordValue.(int))
-	case string:
-		strVal = recordValue.(string)
-	default:
-		strVal = ""
-	}
-	if strVal != "" && strVal != "0" {
-		records = append(records, recordPrepend+": "+strVal)
-	}
-
-	return records
-}
-
-func userIDList(mems []vinyldns.User) string {
-	members := []string{}
-
-	for _, m := range mems {
-		members = append(members, m.ID)
-	}
-
-	return strings.Join(members, ", ")
-}
-
-func printUsers(users []vinyldns.User) {
-	for _, u := range users {
-		data := [][]string{
-			{"UserName", u.UserName},
-			{"Name", u.FirstName + " " + u.LastName},
-			{"ID", u.ID},
-			{"Email", u.Email},
-			{"Created", u.Created},
-		}
-
-		printBasicTable(data)
-	}
-}
-
-func printGroup(group vinyldns.Group) {
-	data := [][]string{
-		{"Name", group.Name},
-		{"Status", group.Status},
-		{"Created", group.Created},
-		{"ID", group.ID},
-	}
-
-	printBasicTable(data)
-}
-
 func printBasicTable(data [][]string) {
 	table := tablewriter.NewWriter(os.Stdout)
 	table.AppendBulk(data)
@@ -219,20 +79,6 @@ func printTableWithHeaders(headers []string, data [][]string) {
 	table.AppendBulk(data)
 	table.SetRowLine(true)
 	table.Render()
-}
-
-func validateConnection(connection string, c *vinyldns.ZoneConnection) (bool, error) {
-	// if all are empty, we assume the user does not want to declare a connection
-	if c.Key == "" && c.KeyName == "" && c.Name == "" && c.PrimaryServer == "" {
-		return false, nil
-	}
-
-	// if any but not all are empty, we have a problem
-	if c.Key == "" || c.KeyName == "" || c.Name == "" || c.PrimaryServer == "" {
-		return false, fmt.Errorf("%s connection requires '--%s-connection-key-name', '--%s-connection-key', and '--%s-connection-primary-server'", connection, connection, connection, connection)
-	}
-
-	return true, nil
 }
 
 func printJSON(i interface{}) error {
